Cache mapping from zero alias to canonical address
The canonical address is assumed to be the first observed address of the account as the first operation referincing the account is it's creation.

See the commit's code comments for explanations of why/when this feature is necessary and why it probably isn't a problem that the above assumption doesn't entirely hold.